Store history on board and graph
And implement Undo functionality

Quick first go:
  - Board stores series of moves and uses these to decide what state to
    change on undo click
  - Graph stores complete state of graph at that point in time

Approach used for board should consume less memory but will take more
effort to implement for the graph as we would need to determine which
connections to re-add. Could do a combination where the edges are stored
and other data is not, could also re-calculate the graph if all edges
and wall positions are known. Will go with this for now and look into
the performance when necessary (e.g. if I ever have a proper go at an
AI).

tidy up